feat(voting): Form validation
